+++ libraries/DataFlash/DataFlash_MAVLink.h
@@ -157,7 +157,7 @@ private:
     void Log_Write_DF_MAV(DataFlash_MAVLink &df);
     
     void internal_error();
-    uint16_t bufferspace_available() override; // in bytes
+    uint32_t bufferspace_available() override; // in bytes
     uint8_t remaining_space_in_current_block();
     // write buffer
     uint8_t _blockcount_free;
